vaccine-scheduler/requirements.txt
# Django and REST Framework
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# Server
gunicorn>=21.0
whitenoise>=6.6

# LangChain and AI
langchain==0.3.7
langchain-community==0.3.7
langchain-core>=0.3.27
langchain-google-genai==2.0.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.2

# Vector DB
faiss-cpu==1.7.4

# Utilities
python-dotenv==1.0.1
pypdf==4.0.1
numpy==1.26.4

# Email Service
resend>=2.0.0
reportlab>=4.0.0
